% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIF.R
\name{CIF}
\alias{CIF}
\title{Estimation of the Cumulative Incidence Function. Gorfine's method,
as discribed by... and A 
Given T1, T2, C, R
`CIF()` return the estimated CIF from the imported data.}
\usage{
CIF(T1, T2, C, R, S_LT, BW_method, Boot = 100, file.name)
}
\arguments{
\item{T1}{The vector that represents the disease age for each participant.
The length of `T1` should be as the length of `T2`,`C`, `R`.
Expected to be numeric vector.}

\item{T2}{The vector that represents the death age for each participant.
The length of `T2` should be as the length of `T1`,`C`, `R`.
Expected to be numeric vector.}

\item{C}{The vector that represents the censor age for each participant.
The length of `C` should be as the length of `T2`,`T1`, `R`.
Expected to be numeric vector.}

\item{R}{The vector that represents the recruited age for each participant.
The length of `R` should be as the length of `T2`,`T1`, `C`.
Expected to be numeric vector.}

\item{S_LT}{The survival to live until the left truncation age, LT.
The default calculated base on the data that the UK government published.
 @seealso https://en.wikipedia.org/wiki/Kernel_(statistics)#cite_note-4}

\item{Boot}{The number of iterations in the bootstrap method. Default set to 100}

\item{kernel}{The kernel function that will be used in "PE" method.
Expected to be one of the follow: CosK(x) EpaK(x) Epa2K(x) gaussK(x),from locpol package.
the default set to 1.}
}
\value{
list object that contains three objects: 'AJ_CIF' contains the CIF that estimated according to Aalen-Johansen method.
  'PE_CIF' contains the CIF that estimated according to MG method.
  'timestocheck' contains the responds ages.
}
\description{
Estimation of the Cumulative Incidence Function. Gorfine's method,
as discribed by... and A 
Given T1, T2, C, R
`CIF()` return the estimated CIF from the imported data.
}
