% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{CIF_new_assumption_checking}
\alias{CIF_new_assumption_checking}
\title{`CIF_new_assumption_checking()` returns assumption test graph, as it describe in ....}
\usage{
CIF_new_assumption_checking(
  T1,
  T2,
  C,
  R,
  S_LT,
  division_val = NULL,
  kernel = locpol::gaussK,
  LT = 40
)
}
\arguments{
\item{T1}{The vector that represents the disease age for each participant.
The length of `T1` should be as the length of `T2`,`C`, `R`.
Expected to be numeric vector.}

\item{T2}{The vector that represents the death age for each participant.
The length of `T2` should be as the length of `T1`,`C`, `R`.
Expected to be numeric vector.}

\item{C}{The vector that represents the censor age for each participant.
The length of `C` should be as the length of `T2`,`T1`, `R`.
Expected to be numeric vector.}

\item{R}{The vector that represents the recruited age for each participant.
The length of `R` should be as the length of `T2`,`T1`, `C`.
Expected to be numeric vector.}

\item{S_LT}{The survival to live until the left truncation age, LT.
The default calculated base on the data that the UK government published.}

\item{division_val}{The gap of years in it the estimator will be examine. Foe example, for division_val=5,
when the range of recruitment is 40-70, every 5 years the estimator performance will be shown.}

\item{kernel}{The kernel function that will be used in "MG" method.
Expected to be one of the follow: 1 - Gaussian, 2 - Epanechnikov,3 - box, 4 - Logistic, 5 - Sigmoid function.
the default set to 1.@seealso https://en.wikipedia.org/wiki/Kernel_(statistics)#cite_note-4}

\item{LT}{The left truncation data of the imported data. If NULL the Left truncation will be the minimum recruited
age in the risk set.}
}
\value{
a list that contains the objects: ggplot object with graph, that contains multiple curves, one for each subgroups;
 Martix that contains the AIC for each object and for whole data.
}
\description{
`CIF_new_assumption_checking()` returns assumption test graph, as it describe in ....
}
