% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_data_base.R
\name{creating_data_base}
\alias{creating_data_base}
\title{Estimation of the Cumulative Incidence Function, CIF by two methods: KM and MG.
that represented on ???
Given T1, T2, C, R.
`creating_base_data()` returns database that contain that needed to the analysis.}
\usage{
creating_data_base(T1, T2, C, R, digits = 0)
}
\arguments{
\item{T1}{The vector that represent the disease age for each participant.
The length of `T1` should be as the length of `T2`,`C`, `R`.
Expected to be numeric vector.}

\item{T2}{The vector that represent the death age for each participant.
The length of `T2` should be as the length of `T1`,`C`, `R`.
Expected to be numeric vector.}

\item{C}{The vector that represent the censor age for each participant.
The length of `C` should be as the length of `T2`,`T1`, `R`.
Expected to be numeric vector.}

\item{R}{The vector that represent the recruited age for each participant.
The length of `C` should be as the length of `T2`,`T1`, `C`.}
}
\value{
data frame that contain:  
 T1,
 T2,
 C,
 V1 = pmin({{T1}},{{T2}},{{C}}),
 V2 = pmin(T2,C),
 delta1 = ifelse(V1==T1 ,1,0),
 delta2 = ifelse(V1==T2,1,0),
 delta3 = ifelse(T2<=C,1,0)*delta1,
 delta4 = ifelse(T2<=C,1,0)
}
\description{
Estimation of the Cumulative Incidence Function, CIF by two methods: KM and MG.
that represented on ???
Given T1, T2, C, R.
`creating_base_data()` returns database that contain that needed to the analysis.
}
