% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_A111}
\alias{dat_A111}
\title{Simulated data set, for  T1=A, T2=1, C=1 and R=1.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{T1}{Age at diagnosis, T1 ~ TWeibull(lambda=4,k=115,a=40)}
  \item{T2}{Age at death. T2 was generated from the empirical distribution of death age in the United Kingdom.
   For observations that were diagnosed with the disease, T1 < T2,
   the original $T_2$ value was discarded, and given T1, a new T2 value was sampled from a truncated Weibull distribution.
   For more details, see in the thesis "Nonparametric cumulative-incidence estimation with delayed entry"}
  \item{R}{Age at recruitment, R ~  U(40,69)}
  \item{C}{Age at censoring, C ~ U(0.1,35)}
 
}
}
\usage{
dat_A111
}
\description{
A simulated dataset containing  5*10^6 rows, that used in the thesis "Nonparametric cumulative-incidence
estimation with delayed entry"
}
\keyword{datasets}
