% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{Graphs}
\alias{Graphs}
\title{Export Graphs of CIF
`Graphs()` function returns plots of the CIF estimators.
 The output is a list og ggplot2 object, hence can be modified by adding ggplot objects using +}
\usage{
Graphs(
  years = NULL,
  CIF_new_sd,
  CIF_new_esti,
  CIF_nelson_sd = NULL,
  CIF_nelson_esti = NULL,
  alpha = 0.05
)
}
\arguments{
\item{years}{Vector of the corspoding years to the estimator,
The length of `T2` should be as the length of `T1`,`C`, `R`.
Expected to be numeric vector. If empty, the years will set to be
years = c(1:length(CIF_new_sd)}

\item{CIF_new_sd}{Vector of the estimator's standard deviation, numeric vector.
The length of CIF_new_sd expected to be equal to CIF_new_esti}

\item{CIF_new_esti}{Vector of the estimator's output, numeric vector.
The length of CIF_new_esti expected to be equal to CIF_new_sd.}
}
\description{
Export Graphs of CIF
`Graphs()` function returns plots of the CIF estimators.
 The output is a list og ggplot2 object, hence can be modified by adding ggplot objects using +
}
