\name{mypac-package}
\alias{mypac-package}
\alias{mypac}
\docType{package}
\title{
\packageTitle{mypac}
}
\description{
\packageDescription{mypac}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mypac}
\packageIndices{mypac}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mypac}

Maintainer: \packageMaintainer{mypac}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
