% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_sim_data.R
\name{sim_data_fun}
\alias{sim_data_fun}
\title{Generate simulated data for the death-illness model
`sim_data_fun()` returns data frame that contains T1,T2,C and R.
 In order to let the user explore the estimators' behavior in
 difference type of diseases and recruitment processes there are multiple options.
 The different type of data allow to explore how the fatality,
 ages' distribution of the eruption, death age and different dependency
 between the disease and the recruitment process affects on the estimators.}
\usage{
sim_data_fun(n, R_method, death_method, T1_method, C_method, LT = 40, U_R = 69)
}
\arguments{
\item{n}{number of observations. The length(n) expected to be 1.}

\item{R_method}{The type of distribution that the vector R will generate from.
 expected to be one of the numbers 0-7.

 Type 0: set all the observations to be R=LT.
 Type 1: generate the observations to from discrete uniform distribution, between LT to U_R.
 Type 2: generate the observations from the UKB's recruitment distribution.
 Type 3-6: generate the observations from normal truncation distribution, between LT to U_R to be
 with dependency upon delta 1:
 1. Type 3
 if the observation i is sick (delta1==1), R_i will arrive from the distribution:

 TN(a=LT, b=U_R, mean = 50, sd = 5),

 otherwise from the distribution:
 TN(a=LT, b=U_R, mean = 55, sd = 10).
 2. Type 4
 if the observation i is sick, R_i will arrive from the distribution:

 TN(a=LT, b=U_R, mean = 55, sd = 10),

 otherwise from the distribution:
 TN(a=LT, b=U_R, mean = 50, sd = 5).

 3. Type 5
 if the observation i is sick, R_i will arrive from the distribuation:

 TN(a=LT, b=U_R, mean = 50, sd = 5),

 otherwise from the distribution:
 TN(a=LT, b=U_R, mean = 60, sd = 7).


 4. Type 6
 if the observation i is sick, R_i will arrive from the distribuation:

 TN(a=LT, b=U_R, mean = 60, sd = 7),

 otherwise from the distribution:
 TN(a=LT, b=U_R, mean = 50, sd = 5).}

\item{death_method}{The death distrbution is base on the UK population.
death_method parmater refers to the disease fatalty. Expect to recive values between 0-4.
Type 0 - The disease not effect on the life expectancy.
Types 1-4 are distrbutions  that represents diffrents disease fatality, with dependenct on the disease age.
All this types base on the truncated normal distbution, when a set to be T1.
1. Type 1: NT(a = T1,b = 105,sd=6,mean=70 )
2. Type 2: NT(a = T1,b = 105,sd=7,mean=50 )
3. Type 3: NT(a = T1,b = 105,sd=7,mean=20 )
4. Type 4: NT(a = T1,b = 105,sd=6,mean=40 )}

\item{T1_method}{Vector of the estimator's output, numeric vector.
The length of CIF_new_esti expected to be equal to CIF_new_sd.}

\item{LT}{The left truncation age of the simulate study,
defult set to be 40 as the UKB data.}

\item{U_R}{The upper adge that observation can get into the simulate study.
defult set to be 69 as the UKB data.}
}
\value{
The function return a data frame with simulate T1,T2,C,R.
}
\description{
Generate simulated data for the death-illness model
`sim_data_fun()` returns data frame that contains T1,T2,C and R.
 In order to let the user explore the estimators' behavior in
 difference type of diseases and recruitment processes there are multiple options.
 The different type of data allow to explore how the fatality,
 ages' distribution of the eruption, death age and different dependency
 between the disease and the recruitment process affects on the estimators.
}
