% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_sim_data.R
\name{sim_data_fun}
\alias{sim_data_fun}
\title{Generate simulated data for the death-illness model
`sim_data_fun()` returns data frame that contains T1,T2,C and R.
To let the user explore the estimators' behavior in
 different types of diseases, recruitment processes, and censoring.}
\usage{
sim_data_fun(n, R_method, death_method, T1_method, C_method, LT = 40, U_R = 69)
}
\arguments{
\item{n}{number of observations. The length(n) expected to be 1.}

\item{R_method}{The type of distribution that the vector R will generate from.
 expected to be one of the numbers 0-2.

 Type 0: set all the observations to be R=LT.
 Type 1: generate the observations to from discrete uniform distribution, between LT to U_R.
 Type 2: generate the observations from the UKB's recruitment distribution.}

\item{death_method}{The death distribution is based on the UK population.
death_method parameter refers to the disease fatality. Expected to values between 0-3.
Type 0 - The disease does not affect life expectancy.
Types 1-3 represent different disease fatalities, with dependence on the disease age.
All these types are based on the truncated Weibull distribution (Tweibull).
1. Type 1: Tweibull(a = T1,b = 105,shape = 5.5 ,scale=55)
2. Type 2: Tweibull(a = T1,b = 105,shape = 5.5 ,scale=68)
3. Type 3: Tweibull(a = T1,b = 105,shape = 5.5 ,scale= 75)}

\item{T1_method}{The disease distribution. Expected to values A-C.
1. Type A: Tweibull(a = 40, shape = 4 ,scale=115)
2. Type B: Tweibull(a = 40, shape = 4 ,scale=130)
3. Type C: Weibull(a = T1, shape=3.5, scale=200)}

\item{LT}{The left truncation age of the simulated study,
default set to be 40 as the UKB data.}

\item{U_R}{The upper edge that observation can get into the simulated study.
default set to be 69 as the UKB data.}
}
\value{
The function returns a data frame with simulated T1, T2, C, R.
}
\description{
Generate simulated data for the death-illness model
`sim_data_fun()` returns data frame that contains T1,T2,C and R.
To let the user explore the estimators' behavior in
 different types of diseases, recruitment processes, and censoring.
}
